Solve day 16
First attempt was too slow to be able to cope with part two. So I did a
couple of different variants to speed things up.

Lesson of the day: If you need to do a lot of string concatenations and
manipulations it may be much faster to work on a list of chars rather
than a string.