Solve day 19.
Lesson of the day: It's important to be aware of the strengths and
shortcommings of the data structures you're using when working with
larger sets of data.

Array list is very slow when removing entries in the beginning.
Linked list is fast when adding and removing items and you don't need to
iterate the list.
Gap list is fast to iterate and fast to delete from if you delete from
the start/end or do all deletions close to each other.