Solve day 21
Fun day. Did a bit of extra work just to create a Scrambler buffer
that do all the scrambling operations in place. Good opportunity for
some TDD practice.

Part two got unexpectedly difficult when it came to reversing
rotation based on a character. Especially when first testing
with the example input and realizing that it's impossible to
reverse 5 character long strings. So in this case you could
only run part 2 on the real data.