Solve day 22.
Part 1 was easy, but part 2 was crazy. Spent a long time trying to figure
out some way of solving it. Couldn't think of anything other than brute
force BFS so started implementing that hoping that the possibilities
would be few enough to handle.

Unfortunately that was not the case. With some hints (solution discussions
on Reddit) I realized that you need to look carefully at your puzzle input
and find a specialized solution that works with the input you have.

When you manage to identify simplifications that can be done to the problem
it's possible to just calculate the number of moves required.

Not very fun with problem where a generic solution is impossible.