Solve day 23 part 2.
Fun day. Fewer optimizations needed than expected. I thought I would
have to understand what the end of the program did and find some
optimizations there as well. But it turned out the initial copying
optimization was enough. I should have tried that as soon as I was
ready with it.