Solve day 24.
This was a tough but interesting day. Obvious what needed to be done
and not too difficult to find a fairly good approach. The problem
ended up being divided in two very separate problems: first parse
the input to some kind of graph, then search trough the graph for
the shortest path.

The first part required a bit of thinking, but after that the
implementation was fairly easy. The graph searching was more tricky
to get correct and find and fix all bugs in the initial implementation.

If you're familiar with A* search that will probably be of great help here,
but I'm not so I just invented my own algorithm.

Lesson of the day: For the X:th time - when doing BFS always check if
a node has already been checked or if it is already in the queue before
enqueueing it.

Still not sure about: Why is my solution 25 times slower when using
a linked list for the queue compared to an array list? There is very
little index based acessing and there are frequent insertions/removals
all over the the queue.