Complete the 2016 event by solving day 25
Very fun problem. Most of the problem was just understanding what the code
does. By adding certain new instructions it was possible to simplify the code
and better understand what it was doing. Most of the code written was unit
tests for verifying that the optimizations and simplificaitons were correct.

Once it was known what the program is doing it was pretty easy to find the
final answer.

One week left until the 2018 event start. Looking forward to it.