2018 Day 5: Alchemical Reduction
It was possible to get a fairly nice and quick solution using a
GapList. Another possibility that also seems pretty nice, but
that I didn't think of is to use a stack for building up a
new string rather than deleting things from the input string.

Part two seemed like a good candidate for trying out coroutines.
Doesn't make much of a difference though since the problem was
small enough (~200ms instead of ~220ms).