2018 Day 6: Chronal Coordinates
This (part 1) was challenging. Finding a way of identifying infinite areas
took quite a bit of thinking and then writing an algorithm that finds
on which side of a diagonal line a given point is located was also tricky.

I ended up doing mostly trial and error with the help of unit tests before
I got it working and understood what I was doing.

Seems like there are much better ways of knowing if an area is infinite
or not. And the puzzle seems to have been small enough that you probably
could have just assumed that areas that are getting too big are infinite.

Part 2 turned out to be more or less trivial and just a few lines of code
that almost didn't reuse any parts of part 1 at all.