2018 Day 11: Chronal Charge
Part 1 was easy and part 2 required thinking and optimizations.

First attempt for part 2 was just the first brute force implementaiton
I could think of. Aborted the execution after about 12 hours.

Second attempt removed some unecessary looping and generations of
tons of lists. This reduced execution time to 11 minutes.

Third attempt was when I realized that grids stretching outside
the 300x300 playground isn't allowed. So when you have a 300x300 grid
there is only 1 such grid. Not 90000 grids where all exept one is
outside the playground. This reduced execution time to 3 minutes

Forth attempt is the current implementation where the sum of one grid
is only calculated once. After that the next grid is calculated as
previous grid - points removed + points added

This reduces the amount of array lookups from n^2 to 2*n and speeds
up the time to solve part 2 to 4 seconds.

Unfortunately I started off with the odd choice of grid[x][y]
addressing, so a couple of things became a bit unintuitive.