2018 Day 12: Subterranean Sustainability
Part 1 was pretty easy, just a bit of fiddling to get the rule
matching correct.

Part 2 was tough. I spotted a pattern early on, but due to minor
off by one bugs it took quite some time to figure out what the
actual pattern is and how to calculate the real value.