2018 Day 14: Chocolate Charts
With an input in the 360k range and knowing that the final list
may be up to twice as long I opted for allocating a byte array
on start and then do all oprations in place to keep things fast.

Probably a good idea for part two (took around 2.5 seconds to solve)
even if it made it a bit more difficult to write the code rather
than working on strings and lists.