2018 Day 19: Go With The Flow
Another assembly problem, fun! Instructions without jumps but instead
directly modifying the instruction pointer with normal instructions
was a bit trickier than usual. So for part 2 it was easier to write
the program in Kotlin to see what it was doing rather than optimizing
the assembler like I did in the 2016 event.