Day 20: A Regular Map
Quite easy when you used a good representation of the data,
but very hard if you do things the wrong way.

First attemt was with a strange kind of tree with strings as
nodes (e.g. EEESSWW) that was completely wrong. Hard to write
the code when the data types used couldn't represent the data
in a correct way.

Second attempt was to not parse the input character by character
but instead with several passes and recursively parse each
parenthesis and put the data in some kind of mix of a tree
and a list. Parsing was slow, the code complicated and it missed
rooms in dead ends (resulted in a off by two answer on my input).

Third attempt was to go back to parsing the input character by
character, but this time use a map to store the location and
distance of each room. Code ended up very easy to read and write
and runs quickly as well.