2018 Day 21: Chronal Conversion
Since the code normally runs forever I had to figure out what it does and especially
when it terminates. I fairly quickly figured out when it checked the given input, but I spent a lot of time after that to figure out what the whole program does.

It was only much later I realized that you could just break when it looks at the input and check what it's looking for.

However, understanding the code was good for part two. I assume there would be a loop somewhere, but I first checked only the first 1000 values and later checked if the first value repeated within the first MAX_INT values which it didnt.

Turned out there were around 3000 non repeating items before the loop started and the loop was almost 10000 items long.