2018 Day 22: Mode Maze
This was pretty tricky. But fun to get some experience with a
bfs with different costs for different moves.

At first I missed that the target coordinates
have special rules. After that there was an additional (small)
challange of re-writing the part 1 solution to be able to walk
outside of the bounding box of (0,0) and the target.

Finally the big challange was to optimize the code to run in a few
seconds rather than minutes. Using a set instead of a list for the
list of checked states was the biggest thing.

Adding things to the queue instead of checking if they are already
there and then just bail out early if it's been checked already
helped a lot as well. Checking if something exists in a set is much
faster than checking if it exists in a list.

Finally adding a heuristic for how things should be prioritized in
the queue helped to cut down the time from 10s to 2.5 seconds.

Still around twice as slow as other kotlin solutions for the same
problem, not sure why.