2018 Day 23: Experimental Emergency Teleportation
Very difficult problem, probably the most difficult AOC problem
for me so far. Hard to know even how to approach it and when
you find an approach it might not work at all due to the large
problem space.

After a long time and several different approaches I found something
that was in the right direction and with some hints I managed to
find a solution that works and that I'm happy with.

https://www.reddit.com/r/adventofcode/comments/b3wneh/need_help_with_2018_day_23_the_nanobot_problem/