2019 Day 5: Sunny with a Chance of Asteroids
More Intcode fun! Not too difficult and only a few small bugs that required some debugging.

Seems like my Intcode design is working fairly well so far. With some refactoring and slight modifications to the day 2 implementation to keep it working I'm pretty happy with how it looks. I'm trying to abstract away the memory reading/wringing logic as well as handling of instructions and parameters so that the actual instruction implementation part can be kept nice and clean.