2019 Day 7: Amplification Circuit
Yay, another Intcode day and a really fun direction with connecting several computers to each other.

Not so difficult even though it required a bit of thinking. Surprisingly and a bit sad, a big part of the challenge (especially for part one) was to write a function that generates permutations of a list of numbers.

My code was more or less correct from the start, but due to copy paste errors in the test I ended up spending some time debugging why the examples for part 2 failed. Turned out I accidentally was running the example code from part 1 and expecting the part 2 output.