2019 Day 8: Space Image Format
I started by solving part 1 by splitting up the data into layers and the layers into rows of pixels. That got complicated and part two got difficult.

So I changed approach and just split it into a bunch of layers as list of ints. But I did the splitting with for loops so it got a bit messy even if it solved the problem. Then I found chunked and everything got simpler, with some thinking I could replace all my loops with standard library components which made everything so much simpler.