2019 Day 9: Sensor Boost
Pretty simple day, but with a lot of changes everywhere. Adding a new parameter mode and instruction was trivial.

Making it possible to address memory outside the original program was easy as
well, just change List<Int> to Map<Int, Int> and modify the overloaded operators
for array access.

Changing all Ints to Longs was not difficult but required a lot of changes everywhere.

Next time there is a similar challenge I need to remember to start out with Longs and Maps from the beginning instead of Ints and Arrays.