2019 Day 10: Monitoring Station
The math was the most difficult part here. In the first part I just calculated the angle
as y/x just to see which asteroids had the same angle.

Turned out that it didn't work well in part 2 where you had to iterate trough the
asteroids based on smallest angle from straight up. Had to spend a lot of time
googling on how trigonometry worked to refresh my memory and realize that
it was some form of "tan" that should be used. Started out with atan and managed
to get something fairly complicated that worked before I learned about atan2
that was much more suitable.

The final challenge was to rotate the coordinate system since the puzzle uses
"y increases downwards" while normal trigonometry is the opposite.

The rest of the puzzle wasn't very hard,  but there were of course some bugs that
had to be solved before everything was working as it should.