2019 Day 12: The N-Body Problem
First part was easy, but the second was really hard. Spent a bit of time
trying to find loops but couldn't discover any fast enough. I didn't
realize that you could assume that the first item would be the first
that repeats.

After I found some loops the next challenge was to implement a
least common multiplier algorithm which I managed after a while. However
it was a simple approach that took way to long to run on large numbers.

Ended up googling for another solution until I found how to implement
it using BigIngeger from Java which was fast enough.

Then it turned out that my loops were incorrect and the next challange
was to figure out why that it wasn't right.

The math was really too hard for me on this puzzle.