2019 Day 14: Space Stoichiometry (partial)
Part one was fairly complicated to come up with an approach and then implement it correctly.

Part two was even more tricky to not introduce off-by-one errors or similar. What
I've done is to run all reactions to produce 1 fuel one by one until a loop is detected
with regards to the amount of leftovers. Then just calculate the amount of fuel
it's possible to store based on that.

However I run out of memory before I can find a loop in the last example and with
my input so I can't complete part two. Hints on Reddit suggests that I should try
to generate more than one fuel at a time. But that is also very slow for me so
that doesn't work.

Looks like my part 1 solution needs to be optimized to do the calculations, possibly
without using recursion to run trough everything.