2019 Day 14: Optimize the material calulation
Don't repeatedly call create with amount 1 when it's possible to call it with a larger
number right away. Makes it take the same amount of time to calculate the amount
of ore needed to create one material as it takes to create large amount of materials.

Should make it possible to solve part 2.