2019 Day 14: Space Stoichiometry (part 2)
There doesn't seem to be any early loops in example 3 and the real input. Now that
it's fast to calculate ore needed for very large amounts of fuel just do a binary
search for the required amount of fuel.

Strangely enough part 2 example 1 fails with this implementation even though all
other examples as well as the real input pass.