2019 Day 16: Flawed Frequency Transmission
It was difficult to understand the description of part 1, but after
that it wasn't too difficult to implement it, even if it was a bit
tricky to get things right.

Part two was completely impossible to me, I couldn't think of any
way of solving it so I had to look for clues on reddit. When you
realized what optimizations and assumptions you could make part
two ended up being easier to implement than part one.