2019 Day 17: Set and Forge
This was a really tough day. Part one was pretty easy, but part two was extreme. It
was very easy to solve it by hand, but I wanted a programmatic solution. The
findPatterns method was the tricky part which took a couple of hours.

This code is just the first that produced the correct output with no refactoring or
cleanup. But now I have something that works, and produces the correct output
which should help later refactoring and cleanup.