2019 Day 20: Donut Maze
Hardest part was to parse the input. Code got pretty messy and it was easy
to do off by one errors. After I solved part 1 I had to re-write the input parsing
part to understand what I was doing. I got fairly nice code, but that was doing too
much filter on the whole map so just the parsing took several seconds instead of
tens of milliseconds.

Had to write something more messy but faster. Part 1 was very easy to solve
once the parsing was done.

Part two was a bit more difficult, the concept was fairly easy but it was a bit tricky
to get everything right. When doing state checking floors have to be considered
but when looking up things in the map the floor should be ignored since all
floors have the same map.

I copied the generic shortest path algorithm and modified instead of updating
the generic one since there where so specific changes needed for part 2.