2019 Day 21: Springdroid Adventure
Nice with an easy problem as a change of pace. It was a bit difficult to understand
what the problem was at first, and especially that the droid's jump distance is 4.

But once I understood the problem it was mostly a matter of finding the correct
boolean logic for when to jump. Part two wasn't that much more difficult either
just a small addition to the part one solution.