2019 Day 23: Category Six
Interesting Intcode problem that required a bit of modifications to my Intcode.

But with the correct modifications part 1 was easy. Part 2 wasn't too difficult either,
but I had some stupid bugs in my NAT code that took a while to find. Then it was a
bit of a challenge to share code between part 1 and 2 in a good way as well.