2019 Day 24: Planet of Discord
Part 1 sounded like a perfect opportunity for bitfields and bitmasks which I used.
Unfortunately there is no Int.foreach that iterates trough the bits in an Integer
so things weren't as smooth as I first hoped. It was a bit tricky to debug as well.

Then I came to part 2 and realized it was completely different and trying to do
bit operations would be a too hard way of finding neighbors. So I did a completely
new approach where I just pre-calculate the neighbors of all the 25 positions
and used that approach instead for both part one and two.