2015 Day 10: Elves Look, Elves Say
First approach was taking a string as input and generate a new string
as output. Part one took a few seconds and part two took 64 minutes.

Changed it to work on list of ints, then part one was in the
millisecond range and part two 2-3 seconds.

Not having to allocate huge amount of strings with many concatenations
does wonders.