2015 Day 11: Corporate Policy
Did a brute force that tests everything and that was pretty quick (less than half a
second for part 2).

It would have been possible to realize that the rules means that the next matching
password would end with the form aabcc and just find the next password with
that pattern.

That would actually have made part 2 slightly more difficult than part 1, now part
2 was just exactly the same as part 1.