2015 Day 12: JSAbacusFramework.io
Part one was very simple. Part two was a bit tricky to get the parsing right, but when
parsing everything into a tree and then sum afterwards things were pretty easy.

Trying to sum as you parse and trying to keep track of whether or not you should
ignore numbers got messy and difficult.