2015 Day 13: Knights of the Dinner Table
I've tried generating all permutations of a list before (2018 day 23 I think) and it was
really difficult for me. So decided to not try to do that again and instead found a
library that does it for me.

Bad news is that the library didn't contain any method for generating necklaces.
Good news is that I found out about Sawada's algorithm that could generate all
necklaces quickly.
Bad news is that the paper would take a huge effort to understand due to too
much math and a programming language I'm not used to.
Good news is that calculating all permutations instead turned out to be fast
enough, so using necklaces was not needed after all.