2015 Day 15: Science for Hungry People
Very difficult to come up with any non brute force solution which I first thought I had
to do. Had to look at others' solutions for clues and realized most people just brute
forced it. 200ms for both parts and the test with brute force. Not at all as slow as
I thought it would be.