2015 Day 20: Infinite Elves and Infinite Houses This was much harder than it had to be.
First I tried to calculate the number of presents for each house
by summing all factors, but that took for ever. Then I spent a lot
of time trying to figure out a way to calculate some lower or
upper bound close to the target to be able to calculate number of
presents from there. But that wasn't really possible either.

Turned out that just simulating the whole thing by having 100k elves
drop presents to 100k houses and finding the first house that matches
the criteria was very quick and easy.