2015 Day 22: Wizard Simulator 20XX
This was surprisingly difficult. Part one took a while to solve
but it went fairly easy even though the code got pretty messy.

In theory part two should be trivial, but it provided the wrong
answer due to several bugs in part 1. I couldn't find all the bugs
so I ended up re-writing the whole thing in a better way which also
required significant debugging before it finally passed part 1.

Then all that was needed for part two was one line of code...
just to realize that there were still bugs in the part 1 implementation.
But this time the code was clean enough to be able to find it (an effect
can give the player more HP, so must make sure the player is not dead
before applying the effect).