2015 Day 25: Let It Snow
A bit of thinking required to figure out how to calculate the index of a given row and
column. But once that was done the problem was small enough to just brute
force calculate the actual code.

Pretty fun year and felt simpler than the last couple of years. Good year to catch up
with after the fact since it didn't require that much time.