2020 Day 4: Passport Processing
Spent quite some time hunting for a bug that caused the real input to generate
an incorrect answer. Turned out that File.readlines skips the last empty
trailing line. It's probably always good, but my solution depended on there
being such a line so I never got the last passport included.

Interesting observation: Making Passport.isValid a property instead of a
function finds the solution in ~1ms instead of ~40ms. Only doing this
calculation once when the passport is created instead of every time
comparing it with others speeds things up a lot.