2020 Day 6: Custom Customs
Finding an initial solution was fairly quick. But after two days with problems
where different entries in the input is separated by double linebreaks I
decided to make a general approach to that. So I'm introducing a new method
for reading input and splitting it on double linebreaks instead of single.