It was a bad idea to put the class to test as a member of the test class
Turns out that the whole test class is instantiated for each test that
is run. The time it takes to instantiate the test class is not measured,
only the time it takes to run the actual test methods. Having a lot
of work done before the actual test method is run just hides the fact
that it's done in the measurements and could result in unnecessary
work being done.

For now it could make sense to let the test input be shared between
all tests. It's not much data and so far the same data has been used
for both parts in most cases.