2020 Day 10: Adapter Array
Part one was really simple, but part two was difficult. I first
tought a bit and came up with solution that allowed for jumps
longer than three in special cases.

So had to spend more time trying to figure out a generic formula
for the amount of combinations given a certain amount of repeating
ones.

I manually calculated the possibilities for 2, 3, and 4 ones trying
to figure out a generic solution. Then I took a look at the input
to see if there were any insanely long sequences thinking that if
there isn't I could probably generate the amount of combinations
instead of coming up with a generic formula.

Then I noticed that the longest sequence that existed in my input
was 4. So no real calculations was actually needed and the problem
trivial since I knew the amount of combinations for 2, 3 and 4 ones
in a row.