2020 Day 13: Shuttle Search
Part 1 was really easy. The same for part two...

if you are familiar with the Chinese Remainder Theorem and have
methods for calculating it, which I wasn't. Figuring out to use
that and then how to implement it was really complicated.

If I remember correctly this method could have been used in day 22
last year. So I've created a generic function for the CRT and
hopefully it'll be able to use it to solve future similar problems
more easily.