2020 Day 15: Rambunctious Recitation
A few off by one errors in part one, then when it was time for part
2 it turned out my solution was good enough from the start so I
just changed the input and could get the solution in a few seconds.

I learned about using sequence {} today which was fun.

Also an interesting observation:
 - a normal MutableMap<Int, Int> solves part 2 in 7.8 seconds
 - a HashMap<Int, Int>(30_000_000) doesn't have to do any
   resizes and solves part 2 in 7.1 seconds
 - an IntArray(30_000_000) doesn't have to do any resizes or
   lookups and solves part 2 in 1.1 seconds.

So an array is much faster than a map when there's a lot of
lookups and the maximum size is known and not a problem.