2020 Day 16: Ticket Translation
First part was easy while the second was a bit trickier.

But several problems caused part two to take some time. First issue
was returning an early debug value in part 2 rather than what the
function calculating the answer returned. The next was to just
multiply indexes instead of the values at the indexes.

The final and biggest problem was the valid ticket detector that
checked if a ticket was valid by checking if the ticket error rate
was non-zero. This would of course treat any invalid ticket where
the only invalid field had value zero as a valid ticket.

The ticket is invalid, but the error rate is also zero.

Other things learned from the reddit solutions megathread:
 * List.iterator() makes it possible to partially iterate
   over a list and continue iterating later. Very useful for
   parsing todays input.
 * The experimental API method buildList makes it possible to
   work with a mutable list and have an immutable version returned.
 * List.associateWith is great for initializing a map with values
   for known keys.