2020 Day 17: Conway Cubes
Like many others I read the description a bit to quickly and missed
"and the frame of view follows the active cells in each cycle" so I
had a very hard time understanding the example when I was trying to
debug my code.

Once I understood that debugging my code was easier and I could spot
my mistakes fairly easy.