2020 Day 20: Jurassic Jigsaw
This was much more fun than the last two days. It was much more
obvious what needed to be done and how to do it.

But there was a lot of code to write and some thinking to do
and of course several bugs to track down so it took quite some
time to get everything working.

Still needs cleanup and improvements, but I'll save that for
another day.