2020 Day 23: Crab Cups
Part one was fairly easy, and I initially used a GapList for fairly
efficient operations which was good enough for part 1. For part
two it was still way to slow so I spent some time trying to calculate
where position 1 would end up but failed.

Had to turn to reddit for hints, where I learned that implementing
a linked list based on arrays is a good idea. So then part 2 became
a fun and manageble challenge.