2020 Day 19: Monster Messages (part 2)
Ugh, this was hard, I couldn't figure out myself what was wrong
and the example for part two was way to large to try to debug
so I had to ask on Reddit for help.

https://www.reddit.com/r/adventofcode/comments/khoho1/2020_day_19_part_2_kotlin_why_cant_i_validate/

Got some really great hints there and a nice minimal test case
that made me realize what was wrong. That also allowed me to find
a very simple test case for part 1 that my implementation failed with.

Knowing what was wrong made it possible to figure out what needed to be
done, but comming up with a working implementation was very tricky.
Took several failed atempts where I mostly tried to use iteration instead
of recursion before I realized how to do it.

In the end the code is fairly simple, especially with all the comments
I put in. But it was quite difficult for me to figure out how get
everything in place. Probably the hardest challenge of this year.