Improved solution for 2020 day 18
Cleaned up the original solution and replaced part 1 with a more
generic version of part 2 that can solve both versions.

To keep the code a bit cleaner several casts are done and assumptions on
which Expressions shows up where. But the input provided is well formed
so this can be safely done. Also leads to execptions being thrown if
bugs are introduced and unexpected types ends up at the wrong place.

Probably still not the nicest solution, but much better than the
first one.