2021 Day 4: Giant Squid
Easy to know what to do, but a bit of work to get it done.

I also had several bugs in the code that took some time to track down
since the test case passed, but the real input failed.

First bug was that I forgot to add the last board to the list of
boards while parsing the input (due to no trailing blank line after
the last board).

Next bug was that I forgot to create a copy of the list I was working
with, so all different boards referenced the same list ending upp with
100 boards identical to the last one in the input file.

With those part 1 succeeded, but part 2 failed. Turned out there was
also a bug with how I detected horizontal bingos so that it could only
detect bingos when the last called number was in the correct column.

So it was just luck that part 1 was accepted. But with that bug
fixed part 2 was also accepted in the end.