2021 Day 7: The Treachery of Whales
Tried to think of something smarter than testing every individual
position, but couldn't think of everything clever. With only a few
thousand positions between min and max a brute force method is fast
enough.

A small optimization was to calculate the fuel cost one position
at a time and stop when cost starts increasing rather than calculating
cost for all positions and take the minimum of that.

A binary search rather than going linary from the start should also
be a bit faster, but the linear approach was fast enough for it to not
be worth implementing a binary search approach.