2021 Day 12: Passage Pathing
I'm not so used with graph traversal so had do think a bit, but
part 1 went fairly well anyway. Part two was a minor change but
required some thinking to get it right with fairly complicated
if statements for whether or not a revisit was allowed.

But once the puzzle was solved and I did a couple of refactorings
the solution ended up pretty nice.

I learned about Map.withDefault today so that was nice!