2021 Day 13: Transparent Origami
This was a fun and fairly easy day. The hardest part in solving
the problem was parsing the input (which wasn't too hard). Folding
was surprisingly easy.

My initial solution had vertical and horizontal folds using almost
the same copy paste code with a few minor tweaks. So the real
challenge was to refactor the code so that the same fold function
could be used for all folds and just pass in different parameters
to it.