2021 Day 14: Extended Polymerization
Part one was easy to solve with brute force. But as expected part
two was impossible with brute force.

Spent some time thinking but couldn't come up with any clever
solution so I had to visit reddit to get some hints. Turns out
the problem was very similar to the lanternfish problem (day 6),
but I failed to realize it.

Once I saw that I kind of understood how to solve the problem and
managed to solve it. But it was a bit more tricky and took a bit
of time to get right.

Interesting how the solution for the lanternfish problem came to
me very easily while I completely failed to see the similarity
with this problem.