Day 18: Snailfish
This took a lot of time. Spent some time thinking about how to
parse and represent the data and went with a simple tree where
each node knew about it's children. But that got really difficult
when doing the expload operation since the nodes had no knowledge
about their parent or siblings.

I gave up on that idea and tried another approach which turned
out to be a complete dead end after a while.

So I gave in and just treated the data as a (linked) list of
characters and modified the list for the various operations.

With everything being just characters it got a bit interesting
when values became larger than 9 since 10 would become ':' there
was also the problem of 43 being '[' which is the start of a
pair and not a large number that needs to be split.

That was fairly easy to solve and after I completed the problem
I refactored the code to use a list of tokens instead of a list
of characters which got rid of the problem completely.

Pretty fun problem, but it took a lot of time.