2021 Day 19: Beacon Scanner
This was tough and reminded me a lot of the nanobots from day 23
2018. But this one was much easier to solve for me than that one.

I had a hard time understanding the description about 24 permutations
so I came up with a solution that didn't need to use that at all and
instead used distances between points.

The code got really messy, but it works. Will have to come back and
refactor it later. This is what I had when I got the accepted answer
without any changes.