2021  Day 20: Trench Map
Seemed like an easy problem, but it was tricky to figure out that
infinity was handled differently in the example inputs and the
real input.

After that I had a real stupid bug where I did
repeat(times) { iteration ->
   ...
   if (times % 2 == 1)
   ...
}

Which of course will behave identically in every iteration. That
took quite some while to track down.

Part two takes almost 2 seconds. Using two grids instead of allocating
a new map every time makes it much faster, but the implementation gets
messy and not so straight forward so keeping this approach instead.