2021 Day 21: Dirac Dice
Part one was really easy, but part two required a significant
amount of thinking to solve. First attempt had some bugs which
was fairly difficult to find since there were no partial results
in the description that you could test against. A huge number
was needed and I got the wrong huge number, so not very easy
to figure out what is wrong or if I'm even on the right way.

But doing some experimentation with games that wins after one
or two turns and only one player made it possible to find the
problems.