2021 Day 22: Reactor Reboot
This was difficult. Working in 3D is challenging and implementing
set operations is hard so this was a struggle for me.

Took a while to figure out how to do subtractions and had to do some
drawing in 1D and 2D first to come up with a good approach.

Doing unions was even worse. When drawing on paper in 2D it was fairly
easy to come up with a way of finding the union but I had real problems
with writing the code for it. Once I finally managed and cleaned up the
code it ended up pretty short.