2021 Day 23: Amphipod
This was horrible and the worst day of the year. Felt like it
should be an easy shortest path problem but I really struggled
with it.

First my solution never completed on the test input (probably due
to bugs), so I rewrote it using arrays instead of AMap and various
classes to make it faster. That also never completed so rewrote
it again but using one array for all the positions rather than
one array for all the amphipods which made the code a bit easier.

This time it was faster and ran out of heap space fairly quickly.
By adding a bunch of unit tests I was able to find a few bugs which
made it pass part 1 example in a couple of seconds but running out
of memory with the real input.

By introducing a heuristic for the distance remaining it made part 1
run in 45 seconds and with some optimizations I got it down to 15
seconds.

Part 2 failed and run out of memory. Only then I realized I used
an array as part of the key for the set of visited states which
is not good. Two different arrays with the same content results
in two entries in a set (where the same kind of lists would result
in only one). Fixing this reduced runtime of part 1 to 300ms and
part 2 to 2 seconds.

Some cleanup and fixing of the unit tests is still needed since
those broke when I quickly hacked in support for part 2.