2021 Day 24: Arithmetic Logic Unit
This was fun even if it took a while to solve.

First part was taking a look at the input and seeing that it was
amost the same code run for each digit and then write kotlin code
that did the same. Then put on a couple of refactorings and
simplifications on that to get a basic understanding of what it
did.

With that in place it was mainly using the debugger and breakpoints
while testing different inputs one digit at a time to figure out
which digits where correct.

I want to come back to this again and clean up the code and maybe
implement an automatic solver for the steps I manually took.