Refactor 2021 day 19 to be faster and more readable
Various changes to make the code more readable and understandable
as well as some optimizations.

Biggest optimization is to keep track of failed alignment attempts
so that those can be avoided in the future as they are already
known to not work.

Another small improvement is to keep the distance between all
beacons in a scanner as a set instead of a list. However this
relies on the iteration order of the set to be the same as items
were added to the set. It works and it makes things a bit faster,
but definitely NSFW.