Introduce a generic Search object for graph traversal
Contains implementations of DFS, Djikstra and A* based on
https://www.redblobgames.com/pathfinding/a-star/introduction.html

That page has a very good and understandable description of the
different algorithms. Interesting to see that my previously home
grown solutions with different cost and heuristics are more or
less the same as Djikstra and A*. So when my solutions has been
too slow has mostly been because I have represented the graph
in a poor way rather than my search algorithm is bad.