Improve 2021 day 15
Use the new Djikstra implementation for path finding instead of
ShortestPath. It wasn't actually ShortestPath that was slow for part 2
it was how the 25 time larger map was generated. The map generation
took almost all the time.

Using just the original map with modulo makes part 2 run in about 1.5
seconds instead of in 45 seconds.