Revert "Make ShortestPath take cost into account"
This change broke both unit tests of ShortestPath and the earlier
puzzle that depended on ShortestPath. Using a priority queue breaks
things from being processed in the order they were added which the
original implementation relied on.

2021 Day 15 uses Search.djikstra instead of shortest path now
so the changes done to shortest path in cdccfcd0 are no longer
needed.

This reverts commit 4b916ddaa7cbe2b01fe004e8c5b13a7cb5abdb74.