Further optimize 2018 day 15
When considering which location to move toward, keep track of
the currently found shortest path. If a position has a longer
manhattan distance than the shortest path it can not be shorter
and can be discarded immediately. If it's in range we can check
how far the path to it is, but make sure to bail out early if it's
longer than the currently shortest path.

This also sorts the initial target positions by distance away from
the current position as the path will likely be shorter to positions
close by. If the first checked path is one of the shortest all other
later paths checks can bail out much earlier compared to if the longest
paths are checked first.

Before these optimizations part 2 took around 8s to solve. With the
optimizations it takes around 700ms - 1 second. Total runtime of
part 1 + 2 (and all extra tests) is about 1 second now.